// src/registrasi/model/enum/status-registrasi.enum.ts
export enum StatusRegistrasi {
    menunggu_pembayaran = "Menunggu Pembayaran",
    menunggu_konfirmasi = "Menunggu Konfirmasi",
    bukti_bayar_ditolak = "Bukti Bayar Ditolak",
    pendaftaran_batal = "Pendaftaran Batal",
    dibayar = "Dibayar",
    lulus = "Lulus",
    lulus_tes_akademik = "Lulus Tes Akademik",
    lulus_bersyarat = "Lulus Bersyarat",
    cadangan = "Cadangan",
    belum_diterima = "Belum Diterima"
}